LayoutTests/media/before-load-member-access-expected.txt

Test that accessing member of a non loaded video works.

RUN(video.played)
RUN(video.buffered)
RUN(video.seekable)
RUN(video.volume)
SUCCESS
END OF TEST
